move adjust logic on transaction end and add tests
