+++ packages/rum-core/test/performance-monitoring/performance-monitoring.spec.js
@@ -296,62 +296,6 @@ describe('PerformanceMonitoring', function() {
     }, 10)
   })
 
-  it('should adjust transaction start based on earliest span start', function(done) {
-    const firstSpan = new Span('first-span-name', 'first-span')
-    firstSpan.end()
-
-    const transaction = new Transaction('/', 'transaction', {})
-    transaction.onEnd = function() {
-      performanceMonitoring.adjustTransactionTime(transaction)
-      expect(transaction._start).toBe(firstSpan._start)
-      expect(transaction._end).toBeGreaterThanOrEqual(lastSpan._end)
-      done()
-    }
-    transaction.spans.push(firstSpan)
-
-    const lastSpan = transaction.startSpan('last-span-name', 'last-span')
-    lastSpan.end()
-
-    transaction.detectFinish()
-  })
-
-  it('should adjust transaction end based on latest span end', function(done) {
-    const transaction = new Transaction('/', 'transaction', {})
-    const transactionStart = transaction._start
-
-    const firstSpan = transaction.startSpan('first-span-name', 'first-span')
-    firstSpan.end()
-
-    const longSpan = transaction.startSpan('long-span-name', 'long-span')
-
-    const lastSpan = transaction.startSpan('last-span-name', 'last-span')
-    lastSpan.end()
-
-    longSpan.end()
-    longSpan.end += 500
-
-    transaction.onEnd = function() {
-      performanceMonitoring.adjustTransactionTime(transaction)
-      expect(transaction._start).toBe(transactionStart)
-      expect(transaction._end).toBeGreaterThanOrEqual(longSpan._end)
-      done()
-    }
-    transaction.detectFinish()
-  })
-
-  it('should truncate active spans after transaction ends', () => {
-    const transaction = new Transaction('transaction', 'transaction')
-    const span = transaction.startSpan('test', 'test')
-    expect(transaction.spans.length).toBe(0)
-    expect(Object.keys(transaction._activeSpans).length).toBe(1)
-    transaction.end()
-
-    performanceMonitoring.adjustTransactionTime(transaction)
-    expect(transaction.spans.length).toBe(1)
-    expect(Object.keys(transaction._activeSpans).length).toBe(0)
-    expect(span.type).toContain('.truncated')
-  })
-
   it('should create correct payload', function() {
     var tr = new Transaction('transaction1', 'transaction1type', {
       transactionSampleRate: 1
@@ -362,9 +306,7 @@ describe('PerformanceMonitoring', function() {
     tr.detectFinish()
 
     expect(tr._end).toBeDefined()
-    if (tr._end === tr._start) {
-      tr._end = tr._end + 100
-    }
+    tr._end = span._end + 100
 
     var payload = performanceMonitoring.createTransactionPayload(tr)
     expect(payload.name).toBe('transaction1')
@@ -374,21 +316,6 @@ describe('PerformanceMonitoring', function() {
     expect(payload.spans[0].type).toBe('span1type')
     expect(payload.spans[0].start).toBe(span._start - tr._start)
     expect(payload.spans[0].duration).toBe(span._end - span._start)
-  })
-
-  it('should not produce negative durations while adjusting to the spans', function() {
-    var transaction = new Transaction('transaction', 'transaction')
-    var span = transaction.startSpan('test', 'test')
-    span.end()
-    span._end += 100
-    span = transaction.startSpan('test', 'external.http')
-
-    span.end()
-    span._start = 10000000
-    span._end = 11000000
-    transaction.end()
-    performanceMonitoring.adjustTransactionTime(transaction)
-    expect(span.duration()).toBe(0)
   })
 
   it('should sendPageLoadMetrics', function(done) {
